Create omitt function and Refactor
